Factor out argv parsing
See #199
